efl_animation: Show end state of animation if duration is 0
If the duration is 0, then the end state of the animation is displayed
and the animation ends.

Use animator to start next member instance in group sequence instance.
Because without animator, the "end" event of the next member instance
happens before the "end" event of the current member instance happens.