elm_test: Remove weight_set(1, 1) from most eo objects
This removes the explicit calls to weight_set(EXPAND) for
most objects created with efl_add(). Still left to do: re-add
the calls to weight_set(0,0) for objects that need no weight
in those examples.

Note: This also rewrites the test case "Ui.Box" to use exclusively
EO APIs. This points out some inconsistencies, but also shows
that the code size can be reduced a bit. I also factorized code.

Note: In the test case Ui.Box the 2 labels simply don't appear as
we could expect. Probably a missing min size and a broken algo in
the box.

Ref T5301